:recycle:refactor(ListCard.jsx): refactor mapping list by making components
List에서 map으로 반복되는 부분을 ListCard component로 쪼개어 리펙토링했습니다.

by @dypark26
